+++ src/renderer/src/app-shell/use-onboarding-and-feature-tips.ts
@@ -1,0 +1,141 @@
+import { useCallback, useEffect, useRef, useState } from 'react'
+import { useShallow } from 'zustand/react/shallow'
+import { onOnboardingReopened } from '../components/onboarding/show-onboarding-event'
+import { shouldShowOnboarding } from '../components/onboarding/should-show-onboarding'
+import {
+  getFeatureTipsAppOpenDecision,
+  isCliFeatureTipCompleted
+} from '../components/feature-tips/feature-tip-startup-gate'
+import {
+  trackCmdJPaletteFeatureTipShown,
+  trackOrcaCliFeatureTipShown
+} from '../components/feature-tips/feature-tip-telemetry'
+import { useAppStore } from '../store'
+import type { OnboardingState } from '../../../shared/onboarding-state-types'
+
+export type OnboardingGate = ReturnType<typeof useOnboardingAndFeatureTips>
+
+/**
+ * Owns first-run education: the onboarding flow's visibility plus the one-per-session
+ * feature-tip prompt, which stays suppressed until onboarding's state is known.
+ */
+export function useOnboardingAndFeatureTips() {
+  const [onboarding, setOnboarding] = useState<OnboardingState | null>(null)
+  const [onboardingLoaded, setOnboardingLoaded] = useState(false)
+  const [featureTipCliInstalled, setFeatureTipCliInstalled] = useState<boolean | null>(null)
+  const promptedThisSessionRef = useRef(false)
+  const suppressedByOnboardingThisSessionRef = useRef(false)
+
+  const activeModal = useAppStore((s) => s.activeModal)
+  const settings = useAppStore((s) => s.settings)
+  const persistedUIReady = useAppStore((s) => s.persistedUIReady)
+  const featureTipsSeenIds = useAppStore((s) => s.featureTipsSeenIds)
+  const featureInteractions = useAppStore((s) => s.featureInteractions)
+  const contextualToursAutoEligible = useAppStore((s) => s.contextualToursAutoEligible)
+  const actions = useAppStore(
+    useShallow((s) => ({
+      openModal: s.openModal,
+      markFeatureTipsSeen: s.markFeatureTipsSeen,
+      setContextualToursAutoEligible: s.setContextualToursAutoEligible,
+      setContextualToursOnboardingVisible: s.setContextualToursOnboardingVisible
+    }))
+  )
+
+  const applyStartupOnboardingState = useCallback((state: OnboardingState): void => {
+    setOnboarding(state)
+    setOnboardingLoaded(true)
+  }, [])
+
+  useEffect(() => {
+    return onOnboardingReopened(setOnboarding)
+  }, [])
+
+  useEffect(() => {
+    // Why: suppress tours until onboarding state is known (null = loading) so a first-run user can't mark a tour seen before onboarding appears.
+    const suppressTours = !onboardingLoaded || shouldShowOnboarding(onboarding)
+    actions.setContextualToursOnboardingVisible(suppressTours)
+  }, [actions, onboarding, onboardingLoaded])
+
+  useEffect(() => {
+    if (!persistedUIReady || !onboardingLoaded || contextualToursAutoEligible !== null) {
+      return
+    }
+    // Why: rollout targets first-run onboarding users; existing profiles are classified once and never auto-toured.
+    actions.setContextualToursAutoEligible(shouldShowOnboarding(onboarding))
+  }, [actions, contextualToursAutoEligible, onboarding, onboardingLoaded, persistedUIReady])
+
+  useEffect(() => {
+    if (!persistedUIReady) {
+      return
+    }
+
+    let cancelled = false
+    void window.api.cli
+      .getInstallStatus()
+      .then((status) => {
+        if (cancelled) {
+          return
+        }
+        setFeatureTipCliInstalled(isCliFeatureTipCompleted(status))
+      })
+      .catch(() => {
+        if (!cancelled) {
+          setFeatureTipCliInstalled(true)
+        }
+      })
+
+    return () => {
+      cancelled = true
+    }
+  }, [persistedUIReady])
+
+  useEffect(() => {
+    const featureTipsDecision = getFeatureTipsAppOpenDecision({
+      activeModal,
+      cliInstalled: featureTipCliInstalled,
+      featureTipsSeenIds,
+      featureInteractions,
+      onboarding,
+      persistedUIReady,
+      promptedThisSession: promptedThisSessionRef.current,
+      settings,
+      suppressedByOnboardingThisSession: suppressedByOnboardingThisSessionRef.current
+    })
+
+    if (featureTipsDecision.kind === 'suppress-for-onboarding') {
+      // Why: first-run users should finish onboarding without a second education modal in the same session.
+      suppressedByOnboardingThisSessionRef.current = true
+      return
+    }
+
+    if (featureTipsDecision.kind !== 'open') {
+      return
+    }
+
+    promptedThisSessionRef.current = true
+    if (featureTipsDecision.tipId === 'orca-cli') {
+      trackOrcaCliFeatureTipShown('app_open')
+    } else if (featureTipsDecision.tipId === 'cmd-j-palette') {
+      trackCmdJPaletteFeatureTipShown('app_open')
+    }
+    // Why: mark seen on show so a quit/crash before dismiss doesn't reappear it next launch.
+    actions.markFeatureTipsSeen([featureTipsDecision.tipId])
+    actions.openModal('feature-tips', { source: 'app_open', tipId: featureTipsDecision.tipId })
+  }, [
+    activeModal,
+    actions,
+    featureTipCliInstalled,
+    featureInteractions,
+    featureTipsSeenIds,
+    onboarding,
+    persistedUIReady,
+    settings
+  ])
+
+  return {
+    applyStartupOnboardingState,
+    onboarding,
+    setOnboarding,
+    shouldRender: onboarding !== null && shouldShowOnboarding(onboarding)
+  }
+}
