+++ src/renderer/src/app-shell/use-runtime-graph-sync.ts
@@ -1,0 +1,60 @@
+import { useEffect } from 'react'
+import { getSystemPrefersDarkSnapshot } from '../components/terminal-pane/use-system-prefers-dark'
+import {
+  canSkipRuntimeMobileSessionSyncKeyBuild,
+  getRuntimeMobileSessionSyncKey,
+  runtimeMobileSessionSyncKeysEqual,
+  scheduleRuntimeGraphSync,
+  setRuntimeGraphStoreStateGetter,
+  setRuntimeGraphSyncEnabled
+} from '../runtime/sync-runtime-graph'
+import { useAppStore } from '../store'
+
+/** Keeps the mobile runtime graph republished as the store's session-visible state changes. */
+export function useRuntimeGraphSync(): void {
+  const workspaceSessionReady = useAppStore((s) => s.workspaceSessionReady)
+
+  useEffect(() => {
+    setRuntimeGraphStoreStateGetter(useAppStore.getState)
+    return () => {
+      setRuntimeGraphStoreStateGetter(null)
+    }
+  }, [])
+
+  useEffect(() => {
+    let previousKey = getRuntimeMobileSessionSyncKey(useAppStore.getState())
+    return useAppStore.subscribe((state, previousState) => {
+      // Why: this fires on every store mutation; read the cached prefers-dark snapshot instead of allocating a throwaway MediaQueryList via matchMedia each tick.
+      const systemPrefersDark = getSystemPrefersDarkSnapshot()
+      // Why: skip the key build when every input is reference-unchanged; the gate mirrors every field getRuntimeMobileSessionSyncKey uses.
+      if (
+        canSkipRuntimeMobileSessionSyncKeyBuild(
+          state,
+          previousState,
+          systemPrefersDark,
+          previousKey.systemPrefersDark
+        )
+      ) {
+        return
+      }
+      const nextKey = getRuntimeMobileSessionSyncKey(
+        state,
+        previousState,
+        previousKey,
+        systemPrefersDark
+      )
+      if (runtimeMobileSessionSyncKeysEqual(nextKey, previousKey)) {
+        return
+      }
+      previousKey = nextKey
+      scheduleRuntimeGraphSync()
+    })
+  }, [])
+
+  useEffect(() => {
+    setRuntimeGraphSyncEnabled(workspaceSessionReady)
+    return () => {
+      setRuntimeGraphSyncEnabled(false)
+    }
+  }, [workspaceSessionReady])
+}
